global, self cancelling idle timer
instead of having a bunch of buffer-local timers polluting the timer
list, we can instead have a single timer that is responsible for
repositioning the overlay.

It checks if the current buffer has window-stool enabled before
repositoning the overlay.

We now also keep track of a list of file-visiting buffers that
have window-stool enabled, so that if all window-stool enabled
buffers are killed, we will self cancel our timer.